success run p2pstore

+++ fastdeploy/entrypoints/openai/serving_chat.py
@@ -708,7 +708,7 @@ class OpenAIServingChat:
                             rd = data["routing_data"]
                             if not isinstance(rd, np.ndarray):
                                 rd = np.array(rd)
-                            routing_data_result = base64.b64encode(rd.astype(np.int32).tobytes()).decode("utf-8")
+                            routing_data_result = base64.b64encode(rd.tobytes()).decode("utf-8")
         finally:
             trace_print(LoggingEventName.POSTPROCESSING_END, request_id, getattr(request, "user", ""))
             tracing.trace_req_finish(request_id)
